# src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Boilerplate",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        }
    ],
    "paths": {
...
        "/api/v1/auth/magic-link/request": {
            "post": {
                "operationId": "v1.auth.magic.request",
                "tags": [
                    "MagicLink"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Always accepted (anti-enumeration).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "accepted": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "accepted"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/admin/v1/orgs/{organization}/restore": {
            "post": {
                "operationId": "admin.v1.orgs.restore",
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/api/v1/profile/password": {
            "put": {
                "operationId": "v1.profile.password",
                "tags": [
                    "Profile"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "changed": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "changed"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/api/v1/auth/2fa/confirm": {
            "post": {
                "operationId": "v1.auth.2fa.confirm",
                "tags": [
                    "TwoFactor"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmTwoFactorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recovery codes \u2014 shown exactly once.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "recovery_codes": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "required": [
                                                "recovery_codes"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/admin/v1/users/{user}/restore": {
            "post": {
                "operationId": "admin.v1.users.restore",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChangePasswordRequest": {
                "type": "object",
                "properties": {
                    "current_password": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "pattern": "[^a-zA-Z0-9]",
                        "minLength": 10,
                        "maxLength": 4096
                    },
                    "password_confirmation": {
                        "type": "string",
                        "pattern": "[^a-zA-Z0-9]",
                        "minLength": 10,
                        "maxLength": 4096
                    }
                },
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "title": "ChangePasswordRequest"
            },
...
            "ConfirmTwoFactorRequest": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "pattern": "^\\d{6}$"
                    }
                },
                "required": [
                    "code"
                ],
                "title": "ConfirmTwoFactorRequest"
            },
...
            "EmailRequest": {
                "type": "object",
                "description": "Body for forgot-password and magic-link/request: just an email.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    }
                },
                "required": [
                    "email"
                ],
                "title": "EmailRequest"
            },
...
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}